fix: gaoc batching order (#3064)
* fix: gaoc batching order

* refactor: apply review suggestions
