+++ src/v0/destinations/mp/util.js
@@ -140,44 +140,6 @@ const isImportAuthCredentialsAvailable = (destination) =>
     destination.Config.projectId);
 
 /**
- * Finds an existing batch based on metadata JobIds from the provided batch and metadataMap.
- * @param {*} batch
- * @param {*} metadataMap The map containing metadata items indexed by JobIds.
- * @returns
- */
-const findExistingBatch = (batch, metadataMap) => {
-  let existingBatch = null;
-
-  // eslint-disable-next-line no-restricted-syntax
-  for (const metadataItem of batch.metadata) {
-    if (metadataMap.has(metadataItem.jobId)) {
-      existingBatch = metadataMap.get(metadataItem.jobId);
-      break;
-    }
-  }
-
-  return existingBatch;
-};
-
-/**
- * Removes duplicate metadata within each merged batch object.
- * @param {*} mergedBatches An array of merged batch objects.
- */
-const removeDuplicateMetadata = (mergedBatches) => {
-  mergedBatches.forEach((batch) => {
-    const metadataSet = new Set();
-    // eslint-disable-next-line no-param-reassign
-    batch.metadata = batch.metadata.filter((metadataItem) => {
-      if (!metadataSet.has(metadataItem.jobId)) {
-        metadataSet.add(metadataItem.jobId);
-        return true;
-      }
-      return false;
-    });
-  });
-};
-
-/**
  * Builds UTM parameters from a campaign object.
  *
  * @param {Object} campaign - The campaign object containing the campaign details.
@@ -274,58 +236,6 @@ const batchEvents = (successRespList, maxBatchSize, reqMetadata) => {
 };
 
 /**
- * Combines batched requests with the same JobIds.
- * @param {*} inputBatches The array of batched request objects.
- * @returns  The combined batched requests with merged JobIds.
- *
- */
-const combineBatchRequestsWithSameJobIds = (inputBatches) => {
-  const combineBatches = (batches) => {
-    const clonedBatches = [...batches];
-    const mergedBatches = [];
-    const metadataMap = new Map();
-
-    clonedBatches.forEach((batch) => {
-      const existingBatch = findExistingBatch(batch, metadataMap);
-
-      if (existingBatch) {
-        // Merge batchedRequests arrays
-        existingBatch.batchedRequest = [
-          ...(Array.isArray(existingBatch.batchedRequest)
-            ? existingBatch.batchedRequest
-            : [existingBatch.batchedRequest]),
-          ...(Array.isArray(batch.batchedRequest) ? batch.batchedRequest : [batch.batchedRequest]),
-        ];
-
-        // Merge metadata
-        batch.metadata.forEach((metadataItem) => {
-          if (!metadataMap.has(metadataItem.jobId)) {
-            metadataMap.set(metadataItem.jobId, existingBatch);
-          }
-          existingBatch.metadata.push(metadataItem);
-        });
-      } else {
-        mergedBatches.push(batch);
-        batch.metadata.forEach((metadataItem) => {
-          metadataMap.set(metadataItem.jobId, batch);
-        });
-      }
-    });
-
-    // Remove duplicate metadata within each merged object
-    removeDuplicateMetadata(mergedBatches);
-
-    return mergedBatches;
-  };
-  // We need to run this twice because in first pass some batches might not get merged
-  // and in second pass they might get merged
-  // Example: [[{jobID:1}, {jobID:2}], [{jobID:3}], [{jobID:1}, {jobID:3}]]
-  // 1st pass: [[{jobID:1}, {jobID:2}, {jobID:3}], [{jobID:3}]]
-  // 2nd pass: [[{jobID:1}, {jobID:2}, {jobID:3}]]
-  return combineBatches(combineBatches(inputBatches));
-};
-
-/**
  * Trims the traits and contextTraits objects based on the setOnceProperties array and returns an object containing the modified traits, contextTraits, and setOnce properties.
  *
  * @param {object} traits - An object representing the traits.
@@ -398,6 +308,5 @@ module.exports = {
   groupEventsByEndpoint,
   generateBatchedPayloadForArray,
   batchEvents,
-  combineBatchRequestsWithSameJobIds,
   trimTraits,
 };
