+++ src/v0/destinations/mp/util.test.js
@@ -1,5 +1,4 @@
 const {
-  combineBatchRequestsWithSameJobIds,
   groupEventsByEndpoint,
   batchEvents,
   generateBatchedPayloadForArray,
@@ -260,235 +259,6 @@ describe('Mixpanel utils test', () => {
       const successRespList = [];
       const result = batchEvents(successRespList, maxBatchSizeMock);
       expect(result).toEqual([]);
-    });
-  });
-
-  describe('Unit test cases for combineBatchRequestsWithSameJobIds', () => {
-    it('Combine batch request with same jobIds', async () => {
-      const input = [
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/track/',
-          },
-          metadata: [
-            {
-              jobId: 1,
-            },
-            {
-              jobId: 4,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/import/',
-          },
-          metadata: [
-            {
-              jobId: 3,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/track/',
-          },
-          metadata: [
-            {
-              jobId: 5,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/engage/',
-          },
-          metadata: [
-            {
-              jobId: 1,
-            },
-            {
-              jobId: 3,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/import/',
-          },
-          metadata: [
-            {
-              jobId: 6,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-      ];
-
-      const expectedOutput = [
-        {
-          batchedRequest: [
-            {
-              endpoint: 'https://api.mixpanel.com/track/',
-            },
-            {
-              endpoint: 'https://api.mixpanel.com/engage/',
-            },
-            {
-              endpoint: 'https://api.mixpanel.com/import/',
-            },
-          ],
-          metadata: [
-            {
-              jobId: 1,
-            },
-            {
-              jobId: 4,
-            },
-            {
-              jobId: 3,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/track/',
-          },
-          metadata: [
-            {
-              jobId: 5,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/import/',
-          },
-          metadata: [
-            {
-              jobId: 6,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-      ];
-      expect(combineBatchRequestsWithSameJobIds(input)).toEqual(expectedOutput);
-    });
-
-    it('Each batchRequest contains unique jobIds (no event multiplexing)', async () => {
-      const input = [
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/track/',
-          },
-          metadata: [
-            {
-              jobId: 1,
-            },
-            {
-              jobId: 4,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/engage/',
-          },
-          metadata: [
-            {
-              jobId: 2,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/engage/',
-          },
-          metadata: [
-            {
-              jobId: 5,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-      ];
-
-      const expectedOutput = [
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/track/',
-          },
-
-          metadata: [
-            {
-              jobId: 1,
-            },
-            {
-              jobId: 4,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/engage/',
-          },
-          metadata: [
-            {
-              jobId: 2,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-        {
-          batchedRequest: {
-            endpoint: 'https://api.mixpanel.com/engage/',
-          },
-          metadata: [
-            {
-              jobId: 5,
-            },
-          ],
-          batched: true,
-          statusCode: 200,
-          destination: destinationMock,
-        },
-      ];
-      expect(combineBatchRequestsWithSameJobIds(input)).toEqual(expectedOutput);
     });
   });
