+++ test/integrations/destinations/google_adwords_offline_conversions/router/data.ts
@@ -96,6 +96,7 @@ export const data = [
                   refresh_token: 'efgh5678',
                   developer_token: 'ijkl91011',
                 },
+                jobId: 1,
                 userId: 'u1',
               },
               destination: {
@@ -211,6 +212,7 @@ export const data = [
                   refresh_token: 'efgh5678',
                   developer_token: 'ijkl91011',
                 },
+                jobId: 2,
                 userId: 'u1',
               },
               destination: {
@@ -274,159 +276,159 @@ export const data = [
                   refresh_token: 'efgh5678',
                   developer_token: 'ijkl91011',
                 },
-                jobId: 1,
-                userId: 'u1',
-              },
-              destination: {
-                Config: {
-                  rudderAccountId: '2Hsy2iFyoG5VLDd9wQcggHLMYFA',
-                  customerId: '769-372-9833',
-                  subAccount: false,
-                  UserIdentifierSource: 'FIRST_PARTY',
-                  conversionEnvironment: 'none',
-                  defaultUserIdentifier: 'email',
-                  hashUserIdentifier: true,
-                  validateOnly: true,
-                  eventsToOfflineConversionsTypeMapping: [
-                    { from: 'Data Reading Guide', to: 'click' },
-                    { from: 'Order Completed', to: 'store' },
-                    { from: 'Sign-up - click', to: 'click' },
-                    { from: 'Outbound click (rudderstack.com)', to: 'click' },
-                    { from: 'Page view', to: 'click' },
-                    { from: 'download', to: 'click' },
-                    { from: 'Product Clicked', to: 'store' },
-                    { from: 'Order Completed', to: 'call' },
-                  ],
-                  loginCustomerId: '4219454086',
-                  eventsToConversionsNamesMapping: [
-                    { from: 'Data Reading Guide', to: 'Data Reading Guide' },
-                    { from: 'Order Completed', to: 'Order Completed' },
-                    { from: 'Sign-up - click', to: 'Sign-up - click' },
-                    {
-                      from: 'Outbound click (rudderstack.com)',
-                      to: 'Outbound click (rudderstack.com)',
-                    },
-                    { from: 'Page view', to: 'Page view' },
-                    { from: 'Sign up completed', to: 'Sign-up - click' },
-                    { from: 'download', to: 'Page view' },
-                    { from: 'Product Clicked', to: 'Store sales' },
-                  ],
-                  authStatus: 'active',
-                  oneTrustCookieCategories: [],
-                  customVariables: [{ from: '', to: '' }],
-                },
-              },
-            },
-            {
-              message: {
-                channel: 'web',
-                context: { traits: { firstName: 'John' } },
-                event: 'Order Completed',
-                type: 'track',
-                messageId: '5e10d13a-bf9a-44bf-b884-43a9e591ea71',
-                originalTimestamp: '2019-10-14T11:15:18.299Z',
-                anonymousId: '00000000000000000000000000',
-                userId: '12345',
-                properties: {
-                  loyaltyFraction: 1,
-                  order_id: 'order id',
-                  currency: 'INR',
-                  revenue: '100',
-                  store_code: 'store code2',
-                  email: 'alex@example.com',
-                  gclid: 'gclid',
-                  product_id: '123445',
-                  quantity: 123,
-                  callerId: '1234',
-                  callStartDateTime: '2019-10-14T11:15:18.299Z',
-                },
-                integrations: { All: true },
-                name: 'ApplicationLoaded',
-                sentAt: '2019-10-14T11:15:53.296Z',
-              },
-              metadata: {
-                secret: {
-                  access_token: 'abcd1234',
-                  refresh_token: 'efgh5678',
-                  developer_token: 'ijkl91011',
-                },
-                jobId: 2,
-                userId: 'u1',
-              },
-              destination: {
-                Config: {
-                  rudderAccountId: '2Hsy2iFyoG5VLDd9wQcggHLMYFA',
-                  customerId: '769-372-9833',
-                  subAccount: false,
-                  UserIdentifierSource: 'FIRST_PARTY',
-                  conversionEnvironment: 'none',
-                  defaultUserIdentifier: 'email',
-                  hashUserIdentifier: true,
-                  validateOnly: true,
-                  eventsToOfflineConversionsTypeMapping: [
-                    { from: 'Data Reading Guide', to: 'click' },
-                    { from: 'Order Completed', to: 'store' },
-                    { from: 'Sign-up - click', to: 'click' },
-                    { from: 'Outbound click (rudderstack.com)', to: 'click' },
-                    { from: 'Page view', to: 'click' },
-                    { from: 'download', to: 'click' },
-                    { from: 'Product Clicked', to: 'store' },
-                    { from: 'Order Completed', to: 'call' },
-                  ],
-                  loginCustomerId: '4219454086',
-                  eventsToConversionsNamesMapping: [
-                    { from: 'Data Reading Guide', to: 'Data Reading Guide' },
-                    { from: 'Order Completed', to: 'Order Completed' },
-                    { from: 'Sign-up - click', to: 'Sign-up - click' },
-                    {
-                      from: 'Outbound click (rudderstack.com)',
-                      to: 'Outbound click (rudderstack.com)',
-                    },
-                    { from: 'Page view', to: 'Page view' },
-                    { from: 'Sign up completed', to: 'Sign-up - click' },
-                    { from: 'download', to: 'Page view' },
-                    { from: 'Product Clicked', to: 'Store sales' },
-                  ],
-                  authStatus: 'active',
-                  oneTrustCookieCategories: [],
-                  customVariables: [{ from: '', to: '' }],
-                },
-              },
-            },
-            {
-              message: {
-                channel: 'web',
-                context: { traits: { firstName: 'John' } },
-                event: 'Order Completed',
-                type: 'track',
-                messageId: '5e10d13a-bf9a-44bf-b884-43a9e591ea71',
-                originalTimestamp: '2019-10-14T11:15:18.299Z',
-                anonymousId: '00000000000000000000000000',
-                userId: '12345',
-                properties: {
-                  loyaltyFraction: 1,
-                  order_id: 'order id',
-                  currency: 'INR',
-                  revenue: '100',
-                  store_code: 'store code2',
-                  email: 'alex@example.com',
-                  gclid: 'gclid',
-                  product_id: '123445',
-                  quantity: 123,
-                  callerId: '1234',
-                  callStartDateTime: '2019-10-14T11:15:18.299Z',
-                },
-                integrations: { All: true },
-                name: 'ApplicationLoaded',
-                sentAt: '2019-10-14T11:15:53.296Z',
-              },
-              metadata: {
-                secret: {
-                  access_token: 'abcd1234',
-                  refresh_token: 'efgh5678',
-                  developer_token: 'ijkl91011',
-                },
                 jobId: 3,
+                userId: 'u1',
+              },
+              destination: {
+                Config: {
+                  rudderAccountId: '2Hsy2iFyoG5VLDd9wQcggHLMYFA',
+                  customerId: '769-372-9833',
+                  subAccount: false,
+                  UserIdentifierSource: 'FIRST_PARTY',
+                  conversionEnvironment: 'none',
+                  defaultUserIdentifier: 'email',
+                  hashUserIdentifier: true,
+                  validateOnly: true,
+                  eventsToOfflineConversionsTypeMapping: [
+                    { from: 'Data Reading Guide', to: 'click' },
+                    { from: 'Order Completed', to: 'store' },
+                    { from: 'Sign-up - click', to: 'click' },
+                    { from: 'Outbound click (rudderstack.com)', to: 'click' },
+                    { from: 'Page view', to: 'click' },
+                    { from: 'download', to: 'click' },
+                    { from: 'Product Clicked', to: 'store' },
+                    { from: 'Order Completed', to: 'call' },
+                  ],
+                  loginCustomerId: '4219454086',
+                  eventsToConversionsNamesMapping: [
+                    { from: 'Data Reading Guide', to: 'Data Reading Guide' },
+                    { from: 'Order Completed', to: 'Order Completed' },
+                    { from: 'Sign-up - click', to: 'Sign-up - click' },
+                    {
+                      from: 'Outbound click (rudderstack.com)',
+                      to: 'Outbound click (rudderstack.com)',
+                    },
+                    { from: 'Page view', to: 'Page view' },
+                    { from: 'Sign up completed', to: 'Sign-up - click' },
+                    { from: 'download', to: 'Page view' },
+                    { from: 'Product Clicked', to: 'Store sales' },
+                  ],
+                  authStatus: 'active',
+                  oneTrustCookieCategories: [],
+                  customVariables: [{ from: '', to: '' }],
+                },
+              },
+            },
+            {
+              message: {
+                channel: 'web',
+                context: { traits: { firstName: 'John' } },
+                event: 'Order Completed',
+                type: 'track',
+                messageId: '5e10d13a-bf9a-44bf-b884-43a9e591ea71',
+                originalTimestamp: '2019-10-14T11:15:18.299Z',
+                anonymousId: '00000000000000000000000000',
+                userId: '12345',
+                properties: {
+                  loyaltyFraction: 1,
+                  order_id: 'order id',
+                  currency: 'INR',
+                  revenue: '100',
+                  store_code: 'store code2',
+                  email: 'alex@example.com',
+                  gclid: 'gclid',
+                  product_id: '123445',
+                  quantity: 123,
+                  callerId: '1234',
+                  callStartDateTime: '2019-10-14T11:15:18.299Z',
+                },
+                integrations: { All: true },
+                name: 'ApplicationLoaded',
+                sentAt: '2019-10-14T11:15:53.296Z',
+              },
+              metadata: {
+                secret: {
+                  access_token: 'abcd1234',
+                  refresh_token: 'efgh5678',
+                  developer_token: 'ijkl91011',
+                },
+                jobId: 4,
+                userId: 'u1',
+              },
+              destination: {
+                Config: {
+                  rudderAccountId: '2Hsy2iFyoG5VLDd9wQcggHLMYFA',
+                  customerId: '769-372-9833',
+                  subAccount: false,
+                  UserIdentifierSource: 'FIRST_PARTY',
+                  conversionEnvironment: 'none',
+                  defaultUserIdentifier: 'email',
+                  hashUserIdentifier: true,
+                  validateOnly: true,
+                  eventsToOfflineConversionsTypeMapping: [
+                    { from: 'Data Reading Guide', to: 'click' },
+                    { from: 'Order Completed', to: 'store' },
+                    { from: 'Sign-up - click', to: 'click' },
+                    { from: 'Outbound click (rudderstack.com)', to: 'click' },
+                    { from: 'Page view', to: 'click' },
+                    { from: 'download', to: 'click' },
+                    { from: 'Product Clicked', to: 'store' },
+                    { from: 'Order Completed', to: 'call' },
+                  ],
+                  loginCustomerId: '4219454086',
+                  eventsToConversionsNamesMapping: [
+                    { from: 'Data Reading Guide', to: 'Data Reading Guide' },
+                    { from: 'Order Completed', to: 'Order Completed' },
+                    { from: 'Sign-up - click', to: 'Sign-up - click' },
+                    {
+                      from: 'Outbound click (rudderstack.com)',
+                      to: 'Outbound click (rudderstack.com)',
+                    },
+                    { from: 'Page view', to: 'Page view' },
+                    { from: 'Sign up completed', to: 'Sign-up - click' },
+                    { from: 'download', to: 'Page view' },
+                    { from: 'Product Clicked', to: 'Store sales' },
+                  ],
+                  authStatus: 'active',
+                  oneTrustCookieCategories: [],
+                  customVariables: [{ from: '', to: '' }],
+                },
+              },
+            },
+            {
+              message: {
+                channel: 'web',
+                context: { traits: { firstName: 'John' } },
+                event: 'Order Completed',
+                type: 'track',
+                messageId: '5e10d13a-bf9a-44bf-b884-43a9e591ea71',
+                originalTimestamp: '2019-10-14T11:15:18.299Z',
+                anonymousId: '00000000000000000000000000',
+                userId: '12345',
+                properties: {
+                  loyaltyFraction: 1,
+                  order_id: 'order id',
+                  currency: 'INR',
+                  revenue: '100',
+                  store_code: 'store code2',
+                  email: 'alex@example.com',
+                  gclid: 'gclid',
+                  product_id: '123445',
+                  quantity: 123,
+                  callerId: '1234',
+                  callStartDateTime: '2019-10-14T11:15:18.299Z',
+                },
+                integrations: { All: true },
+                name: 'ApplicationLoaded',
+                sentAt: '2019-10-14T11:15:53.296Z',
+              },
+              metadata: {
+                secret: {
+                  access_token: 'abcd1234',
+                  refresh_token: 'efgh5678',
+                  developer_token: 'ijkl91011',
+                },
+                jobId: 5,
                 userId: 'u1',
               },
               destination: {
@@ -479,80 +481,130 @@ export const data = [
         body: {
           output: [
             {
-              batchedRequest: {
-                version: '1',
-                type: 'REST',
-                method: 'POST',
-                endpoint:
-                  'https://googleads.googleapis.com/v14/customers/7693729833/offlineUserDataJobs',
-                headers: {
-                  Authorization: 'Bearer abcd1234',
-                  'Content-Type': 'application/json',
-                  'developer-token': 'ijkl91011',
-                },
-                params: { event: 'Store sales', customerId: '7693729833' },
-                body: {
-                  JSON: {
-                    event: '7693729833',
-                    isStoreConversion: true,
-                    createJobPayload: {
-                      job: {
-                        storeSalesMetadata: {
-                          loyaltyFraction: 1,
-                          transaction_upload_fraction: '1',
+              batchedRequest: [
+                {
+                  version: '1',
+                  type: 'REST',
+                  method: 'POST',
+                  endpoint:
+                    'https://googleads.googleapis.com/v14/customers/7693729833/offlineUserDataJobs',
+                  headers: {
+                    Authorization: 'Bearer abcd1234',
+                    'Content-Type': 'application/json',
+                    'developer-token': 'ijkl91011',
+                  },
+                  params: { event: 'Store sales', customerId: '7693729833' },
+                  body: {
+                    JSON: {
+                      event: '7693729833',
+                      isStoreConversion: true,
+                      createJobPayload: {
+                        job: {
+                          storeSalesMetadata: {
+                            loyaltyFraction: 1,
+                            transaction_upload_fraction: '1',
+                          },
+                          type: 'STORE_SALES_UPLOAD_FIRST_PARTY',
                         },
-                        type: 'STORE_SALES_UPLOAD_FIRST_PARTY',
                       },
+                      addConversionPayload: {
+                        operations: [
+                          {
+                            create: {
+                              transaction_attribute: {
+                                store_attribute: { store_code: 'store code' },
+                                transaction_amount_micros: '100000000',
+                                order_id: 'order id',
+                                currency_code: 'INR',
+                                transaction_date_time: '2019-10-14 16:45:18+05:30',
+                              },
+                              userIdentifiers: [
+                                {
+                                  hashedEmail:
+                                    '6db61e6dcbcf2390e4a46af426f26a133a3bee45021422fc7ae86e9136f14110',
+                                },
+                              ],
+                            },
+                          },
+                          {
+                            create: {
+                              transaction_attribute: {
+                                store_attribute: { store_code: 'store code2' },
+                                transaction_amount_micros: '100000000',
+                                order_id: 'order id',
+                                currency_code: 'INR',
+                                transaction_date_time: '2019-10-14 16:45:18+05:30',
+                              },
+                              userIdentifiers: [
+                                {
+                                  hashedEmail:
+                                    '6db61e6dcbcf2390e4a46af426f26a133a3bee45021422fc7ae86e9136f14110',
+                                },
+                              ],
+                            },
+                          },
+                        ],
+                        enable_partial_failure: false,
+                        enable_warnings: false,
+                        validate_only: true,
+                      },
+                      executeJobPayload: { validate_only: true },
                     },
-                    addConversionPayload: {
-                      operations: [
+                    JSON_ARRAY: {},
+                    XML: {},
+                    FORM: {},
+                  },
+                  files: {},
+                },
+                {
+                  version: '1',
+                  type: 'REST',
+                  method: 'POST',
+                  endpoint:
+                    'https://googleads.googleapis.com/v14/customers/7693729833:uploadCallConversions',
+                  headers: {
+                    Authorization: 'Bearer abcd1234',
+                    'Content-Type': 'application/json',
+                    'developer-token': 'ijkl91011',
+                  },
+                  params: {
+                    event: 'Order Completed',
+                    customerId: '7693729833',
+                    customVariables: [{ from: '', to: '' }],
+                    properties: {
+                      loyaltyFraction: 1,
+                      order_id: 'order id',
+                      currency: 'INR',
+                      revenue: '100',
+                      store_code: 'store code2',
+                      email: 'alex@example.com',
+                      gclid: 'gclid',
+                      product_id: '123445',
+                      quantity: 123,
+                      callerId: '1234',
+                      callStartDateTime: '2019-10-14T11:15:18.299Z',
+                    },
+                  },
+                  body: {
+                    JSON: {
+                      conversions: [
                         {
-                          create: {
-                            transaction_attribute: {
-                              store_attribute: { store_code: 'store code' },
-                              transaction_amount_micros: '100000000',
-                              order_id: 'order id',
-                              currency_code: 'INR',
-                              transaction_date_time: '2019-10-14 16:45:18+05:30',
-                            },
-                            userIdentifiers: [
-                              {
-                                hashedEmail:
-                                  '6db61e6dcbcf2390e4a46af426f26a133a3bee45021422fc7ae86e9136f14110',
-                              },
-                            ],
-                          },
-                        },
-                        {
-                          create: {
-                            transaction_attribute: {
-                              store_attribute: { store_code: 'store code2' },
-                              transaction_amount_micros: '100000000',
-                              order_id: 'order id',
-                              currency_code: 'INR',
-                              transaction_date_time: '2019-10-14 16:45:18+05:30',
-                            },
-                            userIdentifiers: [
-                              {
-                                hashedEmail:
-                                  '6db61e6dcbcf2390e4a46af426f26a133a3bee45021422fc7ae86e9136f14110',
-                              },
-                            ],
-                          },
+                          callerId: '1234',
+                          callStartDateTime: '2019-10-14T11:15:18.299Z',
+                          conversionDateTime: '2019-10-14 16:45:18+05:30',
+                          conversionValue: 100,
+                          currencyCode: 'INR',
                         },
                       ],
-                      enable_partial_failure: false,
-                      enable_warnings: false,
-                      validate_only: true,
+                      partialFailure: true,
                     },
-                    executeJobPayload: { validate_only: true },
+                    JSON_ARRAY: {},
+                    XML: {},
+                    FORM: {},
                   },
-                  JSON_ARRAY: {},
-                  XML: {},
-                  FORM: {},
+                  files: {},
                 },
-                files: {},
-              },
+              ],
               metadata: [
                 {
                   secret: {
@@ -560,7 +612,7 @@ export const data = [
                     refresh_token: 'efgh5678',
                     developer_token: 'ijkl91011',
                   },
-                  jobId: 1,
+                  jobId: 3,
                   userId: 'u1',
                 },
                 {
@@ -569,7 +621,7 @@ export const data = [
                     refresh_token: 'efgh5678',
                     developer_token: 'ijkl91011',
                   },
-                  jobId: 2,
+                  jobId: 4,
                   userId: 'u1',
                 },
               ],
@@ -714,6 +766,7 @@ export const data = [
                     refresh_token: 'efgh5678',
                     developer_token: 'ijkl91011',
                   },
+                  jobId: 1,
                   userId: 'u1',
                 },
               ],
@@ -824,6 +877,7 @@ export const data = [
                     refresh_token: 'efgh5678',
                     developer_token: 'ijkl91011',
                   },
+                  jobId: 2,
                   userId: 'u1',
                 },
               ],
@@ -858,54 +912,6 @@ export const data = [
               },
             },
             {
-              batchedRequest: {
-                version: '1',
-                type: 'REST',
-                method: 'POST',
-                endpoint:
-                  'https://googleads.googleapis.com/v14/customers/7693729833:uploadCallConversions',
-                headers: {
-                  Authorization: 'Bearer abcd1234',
-                  'Content-Type': 'application/json',
-                  'developer-token': 'ijkl91011',
-                },
-                params: {
-                  event: 'Order Completed',
-                  customerId: '7693729833',
-                  customVariables: [{ from: '', to: '' }],
-                  properties: {
-                    loyaltyFraction: 1,
-                    order_id: 'order id',
-                    currency: 'INR',
-                    revenue: '100',
-                    store_code: 'store code2',
-                    email: 'alex@example.com',
-                    gclid: 'gclid',
-                    product_id: '123445',
-                    quantity: 123,
-                    callerId: '1234',
-                    callStartDateTime: '2019-10-14T11:15:18.299Z',
-                  },
-                },
-                body: {
-                  JSON: {
-                    conversions: [
-                      {
-                        callerId: '1234',
-                        callStartDateTime: '2019-10-14T11:15:18.299Z',
-                        conversionDateTime: '2019-10-14 16:45:18+05:30',
-                        conversionValue: 100,
-                        currencyCode: 'INR',
-                      },
-                    ],
-                    partialFailure: true,
-                  },
-                  JSON_ARRAY: {},
-                  XML: {},
-                  FORM: {},
-                },
-                files: {},
-              },
               metadata: [
                 {
                   secret: {
@@ -913,61 +919,7 @@ export const data = [
                     refresh_token: 'efgh5678',
                     developer_token: 'ijkl91011',
                   },
-                  jobId: 2,
-                  userId: 'u1',
-                },
-              ],
-              batched: false,
-              statusCode: 200,
-              destination: {
-                Config: {
-                  rudderAccountId: '2Hsy2iFyoG5VLDd9wQcggHLMYFA',
-                  customerId: '769-372-9833',
-                  subAccount: false,
-                  UserIdentifierSource: 'FIRST_PARTY',
-                  conversionEnvironment: 'none',
-                  defaultUserIdentifier: 'email',
-                  hashUserIdentifier: true,
-                  validateOnly: true,
-                  eventsToOfflineConversionsTypeMapping: [
-                    { from: 'Data Reading Guide', to: 'click' },
-                    { from: 'Order Completed', to: 'store' },
-                    { from: 'Sign-up - click', to: 'click' },
-                    { from: 'Outbound click (rudderstack.com)', to: 'click' },
-                    { from: 'Page view', to: 'click' },
-                    { from: 'download', to: 'click' },
-                    { from: 'Product Clicked', to: 'store' },
-                    { from: 'Order Completed', to: 'call' },
-                  ],
-                  loginCustomerId: '4219454086',
-                  eventsToConversionsNamesMapping: [
-                    { from: 'Data Reading Guide', to: 'Data Reading Guide' },
-                    { from: 'Order Completed', to: 'Order Completed' },
-                    { from: 'Sign-up - click', to: 'Sign-up - click' },
-                    {
-                      from: 'Outbound click (rudderstack.com)',
-                      to: 'Outbound click (rudderstack.com)',
-                    },
-                    { from: 'Page view', to: 'Page view' },
-                    { from: 'Sign up completed', to: 'Sign-up - click' },
-                    { from: 'download', to: 'Page view' },
-                    { from: 'Product Clicked', to: 'Store sales' },
-                  ],
-                  authStatus: 'active',
-                  oneTrustCookieCategories: [],
-                  customVariables: [{ from: '', to: '' }],
-                },
-              },
-            },
-            {
-              metadata: [
-                {
-                  secret: {
-                    access_token: 'abcd1234',
-                    refresh_token: 'efgh5678',
-                    developer_token: 'ijkl91011',
-                  },
-                  jobId: 3,
+                  jobId: 5,
                   userId: 'u1',
                 },
               ],
